transparent navbar update
Changes in navbar's background color, from #141d26 to transparent. This gives the impression of a less polluted and a cleaner UI. It feels like the screen gets bigger when scrolling down. Changes have been made in both .light and .dark CSS classes.

Additionally, it can be seen that when scrolling down, the navbar's box shadow gives an unpleasant contrast of colors, even though it is barely unnoticeable at first glance.